cambios

+++ inverse/order_manager.py
@@ -39,7 +39,7 @@ class OrderManagerBearish:
             try:
                 self.print_stats()
                 orders = await self.exchange.watch_orders(self.symbol)
-                open_orders = [o for o in orders if o['info'].get('posSide') == 'long']
+                open_orders = [o for o in orders if o['info'].get('posSide') == 'short']
                 if not open_orders:
                     continue
                 for o in open_orders:
@@ -83,7 +83,7 @@ class OrderManagerBearish:
 
     async def create_order(self, side: str, amount: float, price: float):
         try:
-            params = {'posSide': 'long'}
+            params = {'posSide': 'short'}
             resp = await self.exchange.create_order(
                 self.symbol,
                 'limit',
@@ -119,7 +119,7 @@ class OrderManagerBearish:
 
     async def rebalance(self):
         fetchorders = await self.exchange.fetch_open_orders(self.symbol)
-        open_orders = [o for o in fetchorders if o['info'].get('posSide') == 'long']
+        open_orders = [o for o in fetchorders if o['info'].get('posSide') == 'short']
         net_pos = self.total_sells_filled - self.total_buys_filled
 
         sell_orders = [o for o in open_orders if o['side'] == 'sell']
@@ -160,7 +160,7 @@ class OrderManagerBearish:
                 await self.create_order('buy', self.contracts, p)
 
         await asyncio.sleep(0.02)
-        open_orders_final = [o for o in await self.exchange.fetch_open_orders(self.symbol) if o['info'].get('posSide') == 'long']
+        open_orders_final = [o for o in await self.exchange.fetch_open_orders(self.symbol) if o['info'].get('posSide') == 'short']
         total_final = len(open_orders_final)
         if total_final < self.num_orders:
             faltan = self.num_orders - total_final
